Split `tests/client/rpcKeys`
