--- tests/client/service/keysDecrypt.test.ts
@@ -1,94 +1,0 @@
-import type { Host, Port } from '@/network/types';
-import fs from 'fs';
-import path from 'path';
-import os from 'os';
-import Logger, { LogLevel, StreamHandler } from '@matrixai/logger';
-import { Metadata } from '@grpc/grpc-js';
-import { GRPCServer } from '@/grpc';
-import { KeyManager, utils as keysUtils } from '@/keys';
-import {
-  GRPCClientClient,
-  ClientServiceService,
-  utils as clientUtils,
-} from '@/client';
-import keysDecrypt from '@/client/service/keysDecrypt';
-import * as keysPB from '@/proto/js/polykey/v1/keys/keys_pb';
-import * as testUtils from '../../utils';
-
-describe('keysDecrypt', () => {
-  const logger = new Logger('keysDecrypt test', LogLevel.WARN, [
-    new StreamHandler(),
-  ]);
-  const password = 'helloworld';
-  const authenticate = async (metaClient, metaServer = new Metadata()) =>
-    metaServer;
-  let mockedGenerateKeyPair: jest.SpyInstance;
-  let mockedGenerateDeterministicKeyPair: jest.SpyInstance;
-  beforeAll(async () => {
-    const globalKeyPair = await testUtils.setupGlobalKeypair();
-    mockedGenerateKeyPair = jest
-      .spyOn(keysUtils, 'generateKeyPair')
-      .mockResolvedValue(globalKeyPair);
-    mockedGenerateDeterministicKeyPair = jest
-      .spyOn(keysUtils, 'generateDeterministicKeyPair')
-      .mockResolvedValue(globalKeyPair);
-  });
-  afterAll(async () => {
-    mockedGenerateKeyPair.mockRestore();
-    mockedGenerateDeterministicKeyPair.mockRestore();
-  });
-  let dataDir: string;
-  let keyManager: KeyManager;
-  let grpcServer: GRPCServer;
-  let grpcClient: GRPCClientClient;
-  beforeEach(async () => {
-    dataDir = await fs.promises.mkdtemp(
-      path.join(os.tmpdir(), 'polykey-test-'),
-    );
-    const keysPath = path.join(dataDir, 'keys');
-    keyManager = await KeyManager.createKeyManager({
-      password,
-      keysPath,
-      logger,
-    });
-    const clientService = {
-      keysDecrypt: keysDecrypt({
-        authenticate,
-        keyManager,
-      }),
-    };
-    grpcServer = new GRPCServer({ logger });
-    await grpcServer.start({
-      services: [[ClientServiceService, clientService]],
-      host: '127.0.0.1' as Host,
-      port: 0 as Port,
-    });
-    grpcClient = await GRPCClientClient.createGRPCClientClient({
-      nodeId: keyManager.getNodeId(),
-      host: '127.0.0.1' as Host,
-      port: grpcServer.port,
-      logger,
-    });
-  });
-  afterEach(async () => {
-    await grpcClient.destroy();
-    await grpcServer.stop();
-    await keyManager.stop();
-    await fs.promises.rm(dataDir, {
-      force: true,
-      recursive: true,
-    });
-  });
-  test('decrypts data', async () => {
-    const plainText = Buffer.from('abc');
-    const cipherText = await keyManager.encryptWithRootKeyPair(plainText);
-    const request = new keysPB.Crypto();
-    request.setData(cipherText.toString('binary'));
-    const response = await grpcClient.keysDecrypt(
-      request,
-      clientUtils.encodeAuthFromPassword(password),
-    );
-    expect(response).toBeInstanceOf(keysPB.Crypto);
-    expect(response.getData()).toBe('abc');
-  });
-});
