+++ tests/client/service/keysPasswordChange.test.ts
@@ -4,16 +4,15 @@ import path from 'path';
 import os from 'os';
 import Logger, { LogLevel, StreamHandler } from '@matrixai/logger';
 import { Metadata } from '@grpc/grpc-js';
-import { GRPCServer } from '@/grpc';
-import { KeyManager, utils as keysUtils } from '@/keys';
-import {
-  GRPCClientClient,
-  ClientServiceService,
-  utils as clientUtils,
-} from '@/client';
+import KeyManager from '@/keys/KeyManager';
+import GRPCServer from '@/grpc/GRPCServer';
+import GRPCClientClient from '@/client/GRPCClientClient';
 import keysPasswordChange from '@/client/service/keysPasswordChange';
-import * as sessionsPB from '@/proto/js/polykey/v1/sessions/sessions_pb';
+import { ClientServiceService } from '@/proto/js/polykey/v1/client_service_grpc_pb';
 import * as utilsPB from '@/proto/js/polykey/v1/utils/utils_pb';
+import * as sessionsPB from '@/proto/js/polykey/v1/sessions/sessions_pb';
+import * as clientUtils from '@/client/utils/utils';
+import * as keysUtils from '@/keys/utils';
 import * as testUtils from '../../utils';
 
 describe('keysPasswordChange', () => {
@@ -74,7 +73,7 @@ describe('keysPasswordChange', () => {
     grpcClient = await GRPCClientClient.createGRPCClientClient({
       nodeId: keyManager.getNodeId(),
       host: '127.0.0.1' as Host,
-      port: grpcServer.port,
+      port: grpcServer.getPort(),
       logger,
     });
   });
